+++ RFControl.h
@@ -23,7 +23,7 @@ class RFControl
     static bool compressTimings(unsigned int buckets[8], unsigned int *timings, unsigned int timings_size);
     static bool compressTimingsAndSortBuckets(unsigned int buckets[8], unsigned int *timings, unsigned int timings_size);
     static void sendByTimings(int transmitterPin, unsigned int *timings, unsigned int timings_size, unsigned int repeats = 3);
-    static void sendByCompressedTimings(int transmitterPin, unsigned int* buckets, char* compressTimings, unsigned int repeats = 3); 
+    static void sendByCompressedTimings(int transmitterPin, unsigned long* buckets, char* compressTimings, unsigned int repeats = 3); 
     static unsigned int getLastDuration();
 	static bool existNewDuration();
   private:
